Import from Shenandoah tag aarch64-shenandoah-jdk8u121-b14-shenandoah-merge-2017-03-23

+++ src/share/vm/gc_implementation/shenandoah/shenandoahHeapRegion.inline.hpp
@@ -28,7 +28,7 @@
 #include "runtime/atomic.hpp"
 
 inline void ShenandoahHeapRegion::increase_live_data_words(size_t s) {
-  assert (s <= max_jint, "sanity");
+  assert (s <= (size_t)max_jint, "sanity");
   increase_live_data_words((jint)s);
 }
